add spy to sql

+++ app/helper_functions.py
@@ -349,11 +349,13 @@ def send_agents_ghosts(status, agents, ghost, x, y, i, specop):
     msg = ""
     rstatus = RoundStatus.objects.get(id=1)
     if fleet_time < 1 and rstatus.is_running == True:
-	    if agents > 0:
-	        msg = perform_operation(agent_fleet)
-	        main_fleet.agent += agent_fleet.agent
-	        main_fleet.save()
-	        agent_fleet.delete()
+        if agents > 0:
+            msg = perform_operation(agent_fleet)
+            ignore = ["Observe Planet", "Spy Target"]
+            if specop not in ignore:
+                main_fleet.agent += agent_fleet.agent
+                main_fleet.save()
+                agent_fleet.delete()
 
     return msg
     
@@ -391,9 +393,11 @@ def send_ghosts(status, agents, ghost, x, y, i, specop):
         if ghost > 0:
             msg = perform_incantation(ghost_fleet)
             if specop != "Call to Arms" and planet.owner != status:
-                main_fleet.ghost += ghost_fleet.ghost
-                main_fleet.save()
-                ghost_fleet.delete()
+                ignore = ["Survey System"]
+                if specop not in ignore:
+                    main_fleet.ghost += ghost_fleet.ghost
+                    main_fleet.save()
+                    ghost_fleet.delete()
 
     return msg
